infrastructure/testdriver/virtual_sensors.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver virtual sensors methods</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
"use strict";

promise_test(async t => {
  const sensorType = 'gyroscope';
  await test_driver.create_virtual_sensor({type: sensorType});

  const sensor = new Gyroscope();
  t.add_cleanup(async () => {
    sensor.stop();
    await test_driver.remove_virtual_sensor({type: sensorType});
  });
  const watcher = new EventWatcher(t, sensor, ['activate']);

  sensor.start();
  await watcher.wait_for('activate');
}, "Test that virtual sensor can be created and removed.");

promise_test(async t => {
  const sensorType = 'gyroscope';
  await test_driver.create_virtual_sensor({type: sensorType, connected: false});

  const sensor = new Gyroscope();
  t.add_cleanup(async () => {
    sensor.stop();
    await test_driver.remove_virtual_sensor({type: sensorType});
  });

  const watcher = new EventWatcher(t, sensor, ['error']);
  sensor.start();
  const error = await watcher.wait_for('error');
  assert_equals(error.error.name, 'NotReadableError');
}, "Test that unavailable virtual sensor can be created.");

promise_test(async t => {
  const sensorType = 'gyroscope';
  const create_parameters = {
    'type': sensorType,
    'minSamplingFrequency': 6.0
  };

  await test_driver.create_virtual_sensor(create_parameters);

  const sensor = new Gyroscope({frequency: 5.0});
  t.add_cleanup(async () => {
    sensor.stop();
    await test_driver.remove_virtual_sensor({type: sensorType});
  });

  const watcher = new EventWatcher(t, sensor, ['activate', 'reading', 'error']);
  sensor.start();
  await watcher.wait_for('activate');

  const info = await test_driver.get_virtual_sensor_information(
    {'type': sensorType});

  assert_equals(true, info['isReadingData']);
  assert_equals(create_parameters['minSamplingFrequency'],
                info['requestedSamplingFrequency']);
}, "Test that minimum frequency setting works and virtual sensor information\
    can be fetched.");

promise_test(async t => {
  const sensorType = 'accelerometer';
  await test_driver.create_virtual_sensor({type: sensorType});

  const sensor = new Accelerometer();
  t.add_cleanup(async () => {
    sensor.stop();
    await test_driver.remove_virtual_sensor({type: sensorType});
  });

  const watcher = new EventWatcher(t, sensor, ['activate', 'reading', 'error']);
  sensor.start();
  await watcher.wait_for('activate');

  const update_parameters = {
    'type': sensorType,
    'reading': {
        'x': 1.0,
        'y': 2.0,
        'z': 3.0
    }
  };
  test_driver.update_virtual_sensor(update_parameters);

  await watcher.wait_for('reading');
  assert_equals(sensor.x, update_parameters.reading.x);
  assert_equals(sensor.y, update_parameters.reading.y);
  assert_equals(sensor.z, update_parameters.reading.z);

}, "Test that virtual sensor can be updated.");

</script>
